164796850-feature (loginFunctionality): user login functionality    -restructuring the folder structure
[Delivers #164796850]
